Restructured config format
No longer testing a list of containers.  Simply test the container
specified in the config.
