+++ .coveragerc
@@ -1,3 +1,3 @@
 [report]
 fail_under =
-  50
+  48
