+++ README.md
@@ -8,13 +8,12 @@ Box Liner is a tool which runs a [Docker container][1] against a set of
 [Goss][2] validation tests.
 
 
-Intended to test docker container(s) post build and prior to publishing.  The
+Intended to test a docker container post build and prior to publishing.  The
 `boxliner.yml` file should be checked into the project which hosts the
-Dockerfile, or repo which is responsible for building the container(s).
+Dockerfile, or repo which is responsible for building the container.
 
 Box Liner is not intended to test the multi-container interactions.  This is
-better served by a clustering/scheduling framework.  However, it can test
-multiple containers individually.
+better served by a clustering/scheduling framework.
 
 [1]: https://www.docker.com/
 [2]: https://github.com/aelsabbahy/goss
@@ -31,11 +30,10 @@ Create a `boxliner.yml` with the following content.
 
 ```yaml
 ---
-containers:
-  - image: solita/ubuntu-systemd:latest
-    command: /sbin/init
-    goss_file: relative/path/to/test.yml
-    goss_binary: /path/to/goss/binary/goss-linux-amd64
+image: solita/ubuntu-systemd:latest
+command: /sbin/init
+goss_file: relative/path/to/test.yml
+goss_binary: /path/to/goss/binary/goss-linux-amd64
 ```
 
 Create a `test.yml` with the tests to perform.  Reference Goss' full
