+++ TODO.md
@@ -1,5 +1,4 @@
 # TODO
 
 * Allow container options to be passed through CLI.
-* Should we scan multiple containers?
 * Error handling.
