+++ test/test_config.py
@@ -20,35 +20,66 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 
-from boxliner import container
+import os
+
+import pytest
+
+from boxliner import config
 
 
-def test_config_member(config_instance):
-    x = {
-        'containers': [{
-            'image':
-            'solita/ubuntu-systemd:latest',
-            'command':
-            '/sbin/init',
-            'goss_file':
-            './test/test.yml',
-            'goss_binary':
-            '/Users/jodewey/Downloads/goss-linux-amd64',
-        }]
-    }
-
-    assert x == config_instance.config
+@pytest.fixture
+def _data():
+    return """
+image: solita/ubuntu-systemd:latest
+command: /sbin/init
+goss_file: ./test/test.yml
+goss_binary: /Users/jodewey/Downloads/goss-linux-amd64
+"""
 
 
-def test_containers_property(config_instance):
-    assert isinstance(config_instance.containers[0], container.Container)
+@pytest.fixture
+def _args():
+    return {}
 
 
-def test_containers_property_handles_missing_containers_dict(config_instance):
+@pytest.fixture
+def _command_args():
+    return {}
+
+
+@pytest.fixture
+def _instance(_data, _args, _command_args):
+    return config.Config(_data, _args, _command_args)
+
+
+def test_name_property(_instance):
     pass
-    # TODO(retr0h): Schema validate this.
 
 
-def test_get_config(config_instance):
-    # NOTE(retr0h): This is tested via the config member test.
+def test_image_property(_instance):
+    x = 'solita/ubuntu-systemd:latest'
+
+    assert x == _instance.image
+
+
+def test_command_property(_instance):
+    x = '/sbin/init'
+
+    assert x == _instance.command
+
+
+def test_goss_file_property(_instance):
+    x = os.path.abspath('test/test.yml')
+
+    assert x == _instance.goss_file
+
+
+def test_goss_binary_property(_instance):
+    x = '/Users/jodewey/Downloads/goss-linux-amd64'
+
+    assert x == _instance.goss_binary
+
+
+def test_validate(_instance):
     pass
+    #  _instance.validate()
